requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.1
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
iniconfig==1.1.1
itypes==1.2.0
MarkupSafe==2.1.0
ntlm-auth==1.5.0
openapi-codec==1.3.
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyNaCl==1.5.0
pyparsing==3.0.6
pyspnego==0.5.0
pytest==6.2.5
requests==2.27.1
requests-kerberos==0.14.0
requests-ntlm==1.1.0
simplejson==3.17.6
six==1.16.0
smbprotocol==1.9.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
WMI==1.5.1
xmltodict==0.12.0
